requirements.txt
altair==5.4.1
altgraph==0.17.4
attrs==24.2.0
auto-py-to-exe==2.44.1
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Eel==0.17.0
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.53.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mutagen==1.47.0
narwhals==1.6.0
nest-asyncio==1.6.0
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
plotly==5.24.0
protobuf==5.28.0
psutil==6.0.0
pyarrow==17.0.0
pycparser==2.22
pycryptodomex==3.20.0
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-vlc==3.0.20123
pytube==15.0.0
pytube3==9.6.4
pytz==2024.1
pyudev==0.24.3
pywin32-ctypes==0.2.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.8.0
rpds-py==0.20.0
setuptools==71.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
tk==0.1.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
websockets==12.0
Werkzeug==3.0.4
whichcraft==0.6.1
youtube-dl==2021.12.17
yt-dlp==2024.8.6
zipp==3.20.1
zope.event==5.0
zope.interface==6.4.post2
